+++ decider-gatk-genotype-gvcfs/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1 - 2015-09-22
+- [GP-487](https://jira.oicr.on.ca/browse/GP-487)
+    - Add standard call and emit confidence support and default values
+    - Add support for populating the vcf's ID column using dbSNP
+- Update to workflow-gatk-genotype-gvcfs 1.1
 ## 1.0 - 2015-07-24
 - Initial release of the GATK Genotype GVCFs decider
 - Schedules one workflow run for all files that pass the required filter library-template-type and