+++ workflow-gatk-genotype-gvcfs/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1 - 2015-09-22
+- [GP-487](https://jira.oicr.on.ca/browse/GP-487)
+    - Add standard call and emit confidence support and default values
+    - Add support for populating the vcf's ID column using dbSNP
 ## 1.0 - 2015-07-24
 - Initial release of the GATK Genotype GVCFs workflow
 - GATK 3.4-46